+++ energyquantified/metadata/__init__.py
@@ -1,15 +1,20 @@
 from .area import Allocation, Area, Border
 from .curve import Curve, CurveType, DataType
 from .instance import Instance
-from .place import Place, PlaceType
-from .options import Aggregation, Filter
 from .ohlc import (
     OHLCField,
     ContractPeriod,
     ContinuousContract,
     SpecificContract
 )
-
+from .options import Aggregation, Filter
+from .place import Place, PlaceType
+from .subscription import (
+    Subscription,
+    SubscriptionAccess,
+    SubscriptionType,
+    SubscriptionCollectionPerm
+)
 
 __all__ = [
     # Area
@@ -32,5 +37,10 @@ __all__ = [
     "OHLCField",
     "ContractPeriod",
     "ContinuousContract",
-    "SpecificContract"
+    "SpecificContract",
+    # Subscription
+    "Subscription",
+    "SubscriptionAccess",
+    "SubscriptionType",
+    "SubscriptionCollectionPerm"
 ]