+++ CHANGELOG.rst
@@ -12,6 +12,10 @@ dev
 - Add ``unit`` and ``denominator`` attributes to ``Series`` and ``OHLCList``
 - Add support for unit conversion when loading from the API
 - Add ``curve-type`` parameter for ``eq.metadata.curves()``
+- Add ``Subscription``, ``SubscriptionAccess``, ``SubscriptionType`` and
+  ``SubscriptionCollectionPerm`` models
+- Add ``Curve.subscription`` field, providing the user with subscription
+  information for a curve
 
 **Bugfixes**
 
@@ -21,6 +25,7 @@ dev
 **Breaking changes**
 
 - Removed ``Place.area``. Use ``Place.areas`` instead.
+
 
 0.11
 ----
@@ -37,7 +42,6 @@ dev
 
 - Add ``websocket-client`` v1.5.1
 - Upgrade ``requests`` to minimum 2.31 due to security fixes
-
 
 
 0.10.1